+++ libraries/bio/src/unknown.ts
@@ -1,6 +1,10 @@
 import * as grok from 'datagrok-api/grok';
 import * as ui from 'datagrok-api/ui';
 import * as DG from 'datagrok-api/dg';
+import {SeqPalette, SeqPaletteBase} from './seq-palettes';
+import {getMonomerLibHelper} from './monomer-works/monomer-utils';
+import {PolymerType} from './helm/types';
+import {PolymerTypes} from './helm/consts';
 
 /** makes the color less white, makes the transparency effect always perceptible
  * @param {string} color x coordinate.
@@ -19,9 +23,9 @@ function correctColor(color: string | null): string {
   const normR = r / distToBlack;
   const normG = g / distToBlack;
   const normB = b / distToBlack;
-  if (distToBlack > 210) {
+  if (distToBlack > 210)
     return `rgb(${normR * 210},${normG * 210},${normB * 210})`;
-  }
+
   return DG.Color.toRgb(dgColor);
 }
 
@@ -39,14 +43,12 @@ export class StringUtils {
   }
 }
 
-import {SeqPalette, SeqPaletteBase} from './seq-palettes';
-
 export abstract class UnknownSeqPalette implements SeqPalette {
-  public abstract get(m: string): string;
+  public abstract get(m: string, polymerType?: string): string;
 }
 
 export class GrayAllPalette extends UnknownSeqPalette {
-  public get(m: string): string {
+  public get(_m: string, _polymerType?: string): string {
     return '#666666';
   }
 }
@@ -54,13 +56,42 @@ export class GrayAllPalette extends UnknownSeqPalette {
 
 export class UnknownColorPalette extends UnknownSeqPalette {
   public static palette: string[] = UnknownColorPalette.buildPalette();
-
+  // this way is just more futureproof, when we start distinguishing
+  // between different polymer types for coloring of non natural aa's or nucleotides
+  public static customMonomerColors: {[symbol: string]: {[polymerType: string]: string}} = {};
+  public static polymerTypes: string[] = [];
   private static buildPalette(): string[] {
+    getMonomerLibHelper().then((lh) => {
+      lh.awaitLoaded().then(() => {
+        const monLib = lh.getMonomerLib();
+        monLib.onChanged.subscribe(() => {
+          UnknownColorPalette.customMonomerColors = {};
+          UnknownColorPalette.polymerTypes = monLib.getPolymerTypes();
+          for (const polymerType of this.polymerTypes) {
+            const monomerSymbols = monLib.getMonomerSymbolsByType(polymerType as PolymerType);
+            for (const monomerSymbol of monomerSymbols) {
+              const monomer = monLib.getMonomer(polymerType as PolymerType, monomerSymbol);
+              if (monomer?.meta?.colors?.default?.background) {
+                if (!this.customMonomerColors[monomerSymbol])
+                  this.customMonomerColors[monomerSymbol] = {};
+
+                this.customMonomerColors[monomerSymbol][polymerType] =
+                  correctColor(monomer.meta.colors.default.background);
+              }
+            }
+          }
+        });
+      });
+    });
     const res = ([] as string[]).concat(...Object.values(SeqPaletteBase.colourPalette));
     return res;
   }
 
-  public get(m: string): string {
+  public get(m: string, polymerType?: string): string {
+    const colorObj = UnknownColorPalette.customMonomerColors[m];
+    const polType = polymerType ?? PolymerTypes.PEPTIDE;
+    if (colorObj && colorObj[polType])
+      return colorObj[polType];
     const hash: number = StringUtils.hashCode(m);
     const pI = hash % UnknownColorPalette.palette.length;
     return correctColor(UnknownColorPalette.palette[pI]);