+++ packages/Bio/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Bio changelog
 
+## 2.15.2 (2024-09-18)
+
+Add monomer coloring for fasta/separator from monomer library
+
 ## 2.15.1 (2024-09-10)
 
 ### Bug fixes